Fixed validators.check_routes indices bug
Since the stop_times dataframe is sorted by 'trip_id' and 'stop_sequence' the indices are not guaranteed to be ordered. Therefore the code which checked if the last stop of the previous trip had an arrival/departure time returned a wrong index. A check for the last stop of the last trip was also added.